Move constants to utils/constants.js
